Let CodiMD respect Markdown line-breaks by default
CodiMD by default interprets every line-break as a hard break. This is
nice for poetry, but not useful in the majority of cases. The default
Markdown behaviour is to ignore single line-breaks, unless the line they
end ends in two or more spaces:

https://spec.commonmark.org/0.29/#hard-line-break

HackMD/CodiMD made a mistake early in their development by interpreting
every line-break as a hard break. They feel that introducing the proper
behaviour as a default now would break existing documents, so they
suggest adding this YAML preamble to every document as workaround:

```yaml
---
breaks: false
---
```

This sucks if all your documents are expected to follow the common
Markdown behaviour. This commit makes the behaviour specified in the
specs the default, and lets users who wish to use the classic behaviour
override this by setting `breaks: true` themselves.

For users expecting the proper line-break behaviour, this commit means
that no extra configuration is necessary to get that.

Fixes hackmdio/codimd#1212